# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API proyectoUTM",
        "description": "<html><body>Esta es la documentación de la API para el sistema de Investigación Docencia.<p>Para poder usar esta API es necesario solicitar el token y con esto poder utilizar los servicios proporcionados.</p><br><p><b>Autenticación y token de API</b></p><p>La Universidad Tecnológica de la Mixteca (UTM) proporciona el api-token. Una vez que haya creado y confirmado su cuenta.</p></body></html>",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3000",
    "basePath": "",
    "tags": [
        {
            "name": "Institutos",
            "description": "API para obtener información de los institutos de la UTM"
        },
        {
            "name": "Carreras",
            "description": "API para obtener información de las carreras de la UTM"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/institutos": {
            "get": {
                "tags": [
                    "Institutos"
                ],
                "summary": "Listado de todos los institutos que existen en la UTM.",
                "description": "<p>Servicio que devuelve la información detallada de todos los institutos que existen en la UTM.</p><p>&nbsp;</p><p><em><strong>¿Cuáles son los institutos existentes en la UTM?</strong></em></p>",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "idInstituto": 1,
                                    "codigo": "001",
                                    "nombre": "Instituto de electrónica y mecatrónica"
                                },
                                {
                                    "idInstituto": 2,
                                    "codigo": "002",
                                    "nombre": "Instituto de computación"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Institutos"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Listado de todas las carreras que existen en la UTM.",
                "description": "<p>Servicio que devuelve la información detallada de todas las carreras que existen en la UTM.</p><p>&nbsp;</p><p><em><strong>¿Cuáles son las carreras existentes en la UTM?</strong></em></p>",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "idCarrera": 1,
                                    "codigo": "001",
                                    "nombre": "Ingeniería en computación"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Carrera"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/institutos/{idInstituto}": {
            "get": {
                "tags": [
                    "Institutos"
                ],
                "summary": "Listado de un instituto en particular.",
                "description": "<p>Servicio ...</p>",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "idInstituto": 2,
                                    "codigo": "002",
                                    "nombreInstituto": "Instituto de computación"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Instituto"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "required": true,
                        "description": "idInstituto del instituto del cuál se requiere la información",
                        "type": "number",
                        "default": "8"
                    }
                ]
            }
        },
        "/api/carreras/{idCarrera}": {
            "get": {
                "tags": [
                    "Carreras"
                ],
                "summary": "Listado de una carrera en particular.",
                "description": "<p>Servicio ...</p>",
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": [
                                {
                                    "idCarrera": 2,
                                    "codigo": "002",
                                    "nombreCarrera": "Carrera de computación",
                                    "idInstituto":"001"
                                }
                            ]
                        },
                        "schema": {
                            "$ref": "#/definitions/Carrera"
                        }
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                },
                "parameters": [
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "required": true,
                        "description": "idCarrera de la carrera de la cuál se requiere la información",
                        "type": "number",
                        "default": "8"
                    }
                ]
            }
        },
        "/api/institutos/eliminar/{idInstituto}": {
            "delete": {
                "summary": "Eliminar un instituto",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras/eliminar/{idCarrera}": {
            "delete": {
                "summary": "Eliminar una carrera",
                "operationId": "eliminar",
                "parameters": [
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "id de la carrera a eliminar",
                        "required": true,
                        "default": 14
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/institutos/create": {
            "post": {
                "summary": "Creando un instituto",
                "operationId": "create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "instituto",
                        "in": "body",
                        "description": "Instituto agregado",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NuevoInstituto"
                        }
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras/create": {
            "post": {
                "summary": "Creando una carrera",
                "operationId": "create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "carrera",
                        "in": "body",
                        "description": "carrera agregada",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/NuevaCarrera"
                        }
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/institutos/actualizar/{idInstituto}": {
            "put": {
                "summary": "Modificando un instituto",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idInstituto",
                        "in": "path",
                        "description": "id del Instituto a eliminar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name": "Instituto",
                        "in": "body",
                        "description": "Instituto a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejoInstituto"
                        }
                    }
                ],
                "tags": [
                    "Institutos"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        },
        "/api/carreras/actualizar/{idCarrera}": {
            "put": {
                "summary": "Modificando una carrera",
                "operationId": "actualizar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "idCarrera",
                        "in": "path",
                        "description": "id de la carrera a actualizar",
                        "required": true,
                        "default": 18
                    },
                    {
                        "name":  "carrera",
                        "in": "body",
                        "description":  "carrera a modificar",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViejaCarrera"
                        }
                    }
                ],
                "tags": [
                    "Carreras"
                ],
                "responses": {
                    "201": {
                        "description": "respuesta vacía"
                    },
                    "403": {
                        "description": "Prohibido"
                    }
                }
            }
        }
    },
    "definitions": {
        "Instituto": {
            "type": "object",
            "properties": {
                "idInstituto": {
                    "type": "number",
                    "description": "Identificador del instituto"
                },
                "codigo": {
                    "type": "string",
                    "description": "Código del instituto"
                },
                "nombreInstituto": {
                    "type": "string",
                    "description": "Nombre del instituto"
                }
            }
        },
        "Carrera": {
            "type": "object",
            "properties": {
                "idCarrera": {
                    "type": "number",
                    "description": "Identificador de la carrera"
                },
                "codigoCarrera": {
                    "type": "string",
                    "description": "Código de la carrera"
                },
                "nombre": {
                    "type": "string",
                    "description": "Nombre de la carrera"
                },
                "idInstituto": {
                    "type": "number",
                    "description": "Clave foranea del instituto"
                }
            }
        },
        "ViejoInstituto": {
            "type": "object",
            "required": [
                "codigo",
                "nombre"
            ],
            "properties": {
                "codigo": {
                    "type": "string",
                    "default": "111"
                },
                "nombre": {
                    "type": "string",
                    "default": "Instituto patito"
                }
            }
        },
        "ViejaCarrera": {
            "type": "object",
            "required": [
                "codigoCarrera",
                "nombre",
                "idInstituto"
            ],
            "properties": {
                "codigoCarrera": {
                    "type": "string",
                    "default": "111"
                },
                "nombre": {
                    "type": "string",
                    "default": "Carrera patito"
                },
                "idInstituto": {
                    "type": "string",
                    "default": "002"
                }
            }
        },
        "Institutos": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/Instituto"
            }
        },
        "Carreras": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/Carrera"
            }
        },
        "NuevoInstituto": {
            "type": "object",
            "required": [
                "codigo",
                "nombre"
            ],
            "properties": {
                "codigo": {
                    "type": "string",
                    "default": "111"
                },
                "nombre": {
                    "type": "string",
                    "default": "Instituto patito"
                }
            }
        },
        "NuevaCarrera": {
            "type": "object",
            "required": [
                "codigo",
                "nombre"
            ],
            "properties": {
                "codigoCarrera": {
                    "type": "string",
                    "default": "111"
                },
                "nombre": {
                    "type": "string",
                    "default": "Carrera patito"
                },
                "idInstituto": {
                    "type": "string",
                    "default": "002"
                }
            }
        }
    }
}